Commit: a59658a <-- Faster load times, at last.

+++ src/SVGame/Entities/Misc/MiscServerModel.h
@@ -85,10 +85,10 @@ private:
     uint32_t precachedNoiseIndex{ 0 };
 
     // The actual frame that this model its animation should start off with.
-    int32_t startFrame{ 0 };
+    float startFrame{ 0 };
 
     // The actual frame that this model its animation should end at.
-    int32_t endFrame{ 0 };
+    float endFrame{ 0 };
 
     // The bounding box its bottom left, this can be custom set in map editors.
     vec3_t boundingBoxMins = { -16, -16, 0 };
